+++ mopub-sample/src/main/java/com/mopub/simpleadsdemo/AdUnitDataSource.java
@@ -8,6 +8,7 @@ import android.database.sqlite.SQLiteDatabase;
 import com.mopub.common.logging.MoPubLog;
 
 import java.util.ArrayList;
+import java.util.HashSet;
 import java.util.List;
 
 import static com.mopub.simpleadsdemo.MoPubSQLiteHelper.COLUMN_AD_TYPE;
@@ -17,8 +18,17 @@ import static com.mopub.simpleadsdemo.MoPubSQLiteHelper.COLUMN_ID;
 import static com.mopub.simpleadsdemo.MoPubSQLiteHelper.COLUMN_USER_GENERATED;
 import static com.mopub.simpleadsdemo.MoPubSQLiteHelper.TABLE_AD_CONFIGURATIONS;
 import static com.mopub.simpleadsdemo.MoPubSampleAdUnit.AdType;
+import static com.mopub.simpleadsdemo.MoPubSampleAdUnit.AdType.BANNER;
+import static com.mopub.simpleadsdemo.MoPubSampleAdUnit.AdType.CUSTOM_NATIVE;
+import static com.mopub.simpleadsdemo.MoPubSampleAdUnit.AdType.INTERSTITIAL;
+import static com.mopub.simpleadsdemo.MoPubSampleAdUnit.AdType.LEADERBOARD;
+import static com.mopub.simpleadsdemo.MoPubSampleAdUnit.AdType.LIST_VIEW;
+import static com.mopub.simpleadsdemo.MoPubSampleAdUnit.AdType.MRECT;
+import static com.mopub.simpleadsdemo.MoPubSampleAdUnit.AdType.RECYCLER_VIEW;
+import static com.mopub.simpleadsdemo.MoPubSampleAdUnit.AdType.REWARDED_VIDEO;
 
 class AdUnitDataSource {
+    private Context mContext;
     private MoPubSQLiteHelper mDatabaseHelper;
     private String[] mAllColumns = {
             COLUMN_ID,
@@ -29,23 +39,41 @@ class AdUnitDataSource {
     };
 
     AdUnitDataSource(final Context context) {
+        mContext = context.getApplicationContext();
         mDatabaseHelper = new MoPubSQLiteHelper(context);
+        populateDefaultSampleAdUnits();
+    }
+
+    MoPubSampleAdUnit createDefaultSampleAdUnit(final MoPubSampleAdUnit sampleAdUnit) {
+        return createSampleAdUnit(sampleAdUnit, false);
     }
 
     MoPubSampleAdUnit createSampleAdUnit(final MoPubSampleAdUnit sampleAdUnit) {
+        return createSampleAdUnit(sampleAdUnit, true);
+    }
+
+    private MoPubSampleAdUnit createSampleAdUnit(final MoPubSampleAdUnit sampleAdUnit,
+                                                 final boolean isUserGenerated) {
         final ContentValues values = new ContentValues();
+        final int userGenerated = isUserGenerated ? 1 : 0;
         values.put(COLUMN_AD_UNIT_ID, sampleAdUnit.getAdUnitId());
         values.put(COLUMN_DESCRIPTION, sampleAdUnit.getDescription());
-        values.put(COLUMN_USER_GENERATED, 1);
+        values.put(COLUMN_USER_GENERATED, userGenerated);
         values.put(COLUMN_AD_TYPE, sampleAdUnit.getFragmentClassName());
+
         final SQLiteDatabase database = mDatabaseHelper.getWritableDatabase();
         final long insertId = database.insert(TABLE_AD_CONFIGURATIONS, null, values);
         final Cursor cursor = database.query(TABLE_AD_CONFIGURATIONS, mAllColumns,
                 COLUMN_ID + " = " + insertId, null, null, null, null);
         cursor.moveToFirst();
+
         final MoPubSampleAdUnit newAdConfiguration = cursorToAdConfiguration(cursor);
         cursor.close();
         database.close();
+
+        if (newAdConfiguration != null) {
+            MoPubLog.d("Ad configuration added with id: " + newAdConfiguration.getId());
+        }
         return newAdConfiguration;
     }
 
@@ -58,8 +86,7 @@ class AdUnitDataSource {
     }
 
     List<MoPubSampleAdUnit> getAllAdUnits() {
-        final List<MoPubSampleAdUnit> adConfigurations =
-                new ArrayList<MoPubSampleAdUnit>();
+        final List<MoPubSampleAdUnit> adConfigurations = new ArrayList<>();
         SQLiteDatabase database = mDatabaseHelper.getReadableDatabase();
         final Cursor cursor = database.query(TABLE_AD_CONFIGURATIONS,
                 mAllColumns, null, null, null, null, null);
@@ -74,6 +101,65 @@ class AdUnitDataSource {
         cursor.close();
         database.close();
         return adConfigurations;
+    }
+
+    List<MoPubSampleAdUnit> getDefaultAdUnits() {
+        final List<MoPubSampleAdUnit> adUnitList = new ArrayList<>();
+        adUnitList.add(
+                new MoPubSampleAdUnit
+                        .Builder(mContext.getString(R.string.ad_unit_id_banner), BANNER)
+                        .description("MoPub Banner Sample")
+                        .build());
+        adUnitList.add(
+                new MoPubSampleAdUnit
+                        .Builder(mContext.getString(R.string.ad_unit_id_mrect), MRECT)
+                        .description("MoPub Mrect Sample")
+                        .build());
+        adUnitList.add(
+                new MoPubSampleAdUnit
+                        .Builder(mContext.getString(R.string.ad_unit_id_leaderboard), LEADERBOARD)
+                        .description("MoPub Leaderboard Sample")
+                        .build());
+        adUnitList.add(
+                new MoPubSampleAdUnit
+                        .Builder(mContext.getString(R.string.ad_unit_id_interstitial), INTERSTITIAL)
+                        .description("MoPub Interstitial Sample")
+                        .build());
+        adUnitList.add(
+                new MoPubSampleAdUnit
+                        .Builder(mContext.getString(R.string.ad_unit_id_rewarded_video),
+                        REWARDED_VIDEO)
+                        .description("MoPub Rewarded Video Sample")
+                        .build());
+        adUnitList.add(
+                new MoPubSampleAdUnit
+                        .Builder(mContext.getString(R.string.ad_unit_id_native), LIST_VIEW)
+                        .description("MoPub Ad Placer Sample")
+                        .build());
+        adUnitList.add(
+                new MoPubSampleAdUnit
+                        .Builder(mContext.getString(R.string.ad_unit_id_native), RECYCLER_VIEW)
+                        .description("MoPub Recycler View Sample")
+                        .build());
+        adUnitList.add(
+                new MoPubSampleAdUnit
+                        .Builder(mContext.getString(R.string.ad_unit_id_native), CUSTOM_NATIVE)
+                        .description("MoPub View Pager Sample")
+                        .build());
+        return adUnitList;
+    }
+
+    private void populateDefaultSampleAdUnits() {
+        final HashSet<MoPubSampleAdUnit> allAdUnits = new HashSet<>();
+        for (final MoPubSampleAdUnit adUnit : getAllAdUnits()) {
+            allAdUnits.add(adUnit);
+        }
+
+        for (final MoPubSampleAdUnit defaultAdUnit : getDefaultAdUnits()) {
+            if (!allAdUnits.contains(defaultAdUnit)) {
+                createDefaultSampleAdUnit(defaultAdUnit);
+            }
+        }
     }
 
     private MoPubSampleAdUnit cursorToAdConfiguration(final Cursor cursor) {