+++ mopub-sample/src/main/java/com/mopub/simpleadsdemo/LeaderboardDetailFragment.java
@@ -1,0 +1,14 @@
+package com.mopub.simpleadsdemo;
+
+public class LeaderboardDetailFragment extends AbstractBannerDetailFragment {
+
+    @Override
+    public int getWidth() {
+        return (int) getResources().getDimension(R.dimen.leaderboard_width);
+    }
+
+    @Override
+    public int getHeight() {
+        return (int) getResources().getDimension(R.dimen.leaderboard_height);
+    }
+}